# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Loan Backend API",
        "contact": {}
    },
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get user list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user list",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "username",
                            "email",
                            "nik",
                            "fullname",
                            "legalname",
                            "role"
                        ],
                        "type": "string",
                        "description": "leave empty",
                        "name": "query_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "updated_at",
                            "username",
                            "email",
                            "nik",
                            "fullname",
                            "legalname",
                            "role"
                        ],
                        "type": "string",
                        "default": "updated_at",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "x-enum-varnames": [
                            "SortOrder_asc",
                            "SortOrder_desc"
                        ],
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetUserListRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user (current user)",
                "parameters": [
                    {
                        "description": "User update request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UpdateUserRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user (admin only)",
                "parameters": [
                    {
                        "description": "User create request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreateUserRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/face-photo": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Upload my Face photo (current user)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Face photo file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UploadFacePhotoRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/ktp-photo": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Upload my KTP photo (current user)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "KTP photo file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UploadKtpPhotoRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user (current user)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetUserByUUIDResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/{uuid}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetUserByUUIDResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user by UUID (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User update request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UpdateUserRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete user by UUID (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseJSONResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DeleteUserRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BaseJSONResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "detail": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateUserReq": {
            "type": "object",
            "required": [
                "birthdate",
                "birthplace",
                "current_salary",
                "email",
                "fullname",
                "legalname",
                "nik",
                "password",
                "role",
                "username"
            ],
            "properties": {
                "birthdate": {
                    "description": "DD-MM-YYYY",
                    "type": "string"
                },
                "birthplace": {
                    "type": "string"
                },
                "current_salary": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "legalname": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "enum": [
                        "admin",
                        "user"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.UserRole"
                        }
                    ]
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.CreateUserRespData": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "birthplace": {
                    "type": "string"
                },
                "current_salary": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "face_photo": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "ktp_photo": {
                    "type": "string"
                },
                "legalname": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.DeleteUserRespData": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "birthplace": {
                    "type": "string"
                },
                "current_salary": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "face_photo": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "ktp_photo": {
                    "type": "string"
                },
                "legalname": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.GetUserByUUIDResp": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "birthplace": {
                    "type": "string"
                },
                "current_salary": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "face_photo": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "ktp_photo": {
                    "type": "string"
                },
                "legalname": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.GetUserListRespData": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BaseUserResponse"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateUserReq": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "description": "DD-MM-YYYY",
                    "type": "string"
                },
                "birthplace": {
                    "type": "string"
                },
                "current_salary": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "legalname": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/enum.UserRole"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateUserRespData": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "birthplace": {
                    "type": "string"
                },
                "current_salary": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "face_photo": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "ktp_photo": {
                    "type": "string"
                },
                "legalname": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.UploadFacePhotoRespData": {
            "type": "object",
            "properties": {
                "face_photo": {
                    "type": "string"
                }
            }
        },
        "dto.UploadKtpPhotoRespData": {
            "type": "object",
            "properties": {
                "ktp_photo": {
                    "type": "string"
                }
            }
        },
...
        "enum.UserRole": {
            "type": "string",
            "enum": [
                "user",
                "admin"
            ],
            "x-enum-varnames": [
                "UserRole_User",
                "UserRole_Admin"
            ]
        },
        "model.BaseUserResponse": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "birthplace": {
                    "type": "string"
                },
                "current_salary": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "face_photo": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "ktp_photo": {
                    "type": "string"
                },
                "legalname": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "description": "JWT Authorization header using the Bearer scheme (add 'Bearer ' prefix).",
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/auth/login/dev"
        }
    }
}